Do not allow nil to be supplied as input
There is at least one weird issue if `nil` is allowed as input: should
we decide to use the default value if `nil` is explicitly passed in? I
can't think of any good reason to allow `nil` as in input, in fact. If
you want `nil` to be a possible value for a dependency, just set the
default value to `nil`.

 * Closes #1